+++ pykern/api/__init__.py
@@ -1,21 +1,6 @@
-"""HTTP WebSocket server & client
+"""HTTP WebSocket Quest server & client
 
 
 :copyright: Copyright (c) 2024 RadiaSoft LLC.  All Rights Reserved.
 :license: http://www.apache.org/licenses/LICENSE-2.0.html
 """
-
-from pykern.pkcollections import PKDict
-from pykern.pkdebug import pkdc, pkdlog, pkdp, pkdexc, pkdformat
-import asyncio
-import inspect
-import msgpack
-import pykern.pkasyncio
-import pykern.pkcollections
-import pykern.pkconfig
-import pykern.quest
-import pykern.util
-import re
-import tornado.httpclient
-import tornado.web
-import tornado.websocket
